requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.7.0
torchvision==0.22.0
Pillow==11.2.1
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
lime==0.2.0.1
scikit-image==0.22.0
matplotlib==3.8.2
streamlit==1.31.1 